rename recipe list page to search
