+++ app/src/main/java/com/me/recipe/ui/search/component/RecipeList.kt
@@ -1,27 +1,26 @@
 @file:OptIn(ExperimentalSharedTransitionApi::class)
 
-package com.me.recipe.ui.search
+package com.me.recipe.ui.search.component
 
 import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionScope
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
+import com.me.recipe.domain.features.recipe.model.Recipe
 import com.me.recipe.ui.component.util.SharedTransitionLayoutPreview
-import com.me.recipe.ui.search.component.RecipeCard
-import com.me.recipe.ui.search.component.SearchContent
+import com.me.recipe.ui.search.SearchContract
 import com.me.recipe.ui.theme.RecipeTheme
+import kotlinx.collections.immutable.ImmutableList
 
 
 @Composable
 internal fun RecipeList(
-    state: SearchContract.State,
+    recipes: ImmutableList<Recipe>,
     event: (SearchContract.Event) -> Unit,
     sharedTransitionScope: SharedTransitionScope,
     animatedVisibilityScope: AnimatedVisibilityScope,
@@ -30,7 +29,7 @@ internal fun RecipeList(
     LazyColumn(
         modifier = modifier.testTag("testTag_recipeList"),
     ) {
-        itemsIndexed(state.recipes) { index, recipe ->
+        itemsIndexed(recipes) { index, recipe ->
             event.invoke(SearchContract.Event.OnChangeRecipeScrollPosition(index))
 
             RecipeCard(
@@ -54,7 +53,7 @@ private fun SearchContentPreview() {
     RecipeTheme(true) {
         SharedTransitionLayoutPreview {
             RecipeList(
-                state = SearchContract.State.testData(),
+                recipes = SearchContract.State.testData().recipes,
                 event = {},
                 sharedTransitionScope = this,
                 animatedVisibilityScope = it,
